+++ apps/docs/src/app/platform/component-docs/platform-menu/platform-menu-examples/platform-menu-cascade-example.component.ts
@@ -1,0 +1,17 @@
+import { Component, OnInit, ViewEncapsulation } from '@angular/core';
+
+@Component({
+    selector: 'fdp-platform-menu-cascade-example',
+    templateUrl: './platform-menu-cascade-example.component.html',
+    styleUrls: ['./platform-menu-basic-example.component.scss'],
+    encapsulation: ViewEncapsulation.None
+})
+export class PlatformMenuCascadeExampleComponent implements OnInit {
+    public item = '';
+
+    ngOnInit() { }
+
+    onItemSelect(item: string) {
+        this.item = item;
+    }
+}