+++ libs/core/src/lib/datetime-picker/datetime-picker.component.ts
@@ -15,7 +15,6 @@ import {
     ViewEncapsulation
 } from '@angular/core';
 import { AbstractControl, ControlValueAccessor, NG_VALIDATORS, NG_VALUE_ACCESSOR, Validator } from '@angular/forms';
-import { Subscription } from 'rxjs';
 import { TimeObject } from '../time/time-object';
 import { TimeComponent } from '../time/time.component';
 import { Placement } from 'popper.js';
@@ -61,11 +60,11 @@ import { DatePipe } from '@angular/common';
 export class DatetimePickerComponent implements OnInit, ControlValueAccessor, Validator {
 
     /** @hidden Reference to the inner time component. */
-    @ViewChild(TimeComponent, { static: false })
+    @ViewChild(TimeComponent)
     timeComponent: TimeComponent;
 
     /** @hidden Reference to the inner calendar component. */
-    @ViewChild(CalendarComponent, { static: false })
+    @ViewChild(CalendarComponent)
     calendarComponent: CalendarComponent;
 
     /**