+++ libs/core/src/lib/nested-list/nested-list-state.service.ts
@@ -1,7 +1,9 @@
+import { Injectable } from '@angular/core';
+import { Subject } from 'rxjs';
 import { NestedListInterface } from './nested-list/nested-list.interface';
 import { NestedItemInterface } from './nested-item/nested-item.interface';
-import { Subject } from 'rxjs';
 
+@Injectable()
 export class NestedListStateService {
 
     /**