+++ libs/core/src/lib/notification/directives/notification-content.directive.spec.ts
@@ -9,7 +9,7 @@ import { NotificationModule } from '@fundamental-ngx/core';
     `
 })
 class TestComponent {
-    @ViewChild('directiveElement', { static: false })
+    @ViewChild('directiveElement')
     ref: ElementRef;
 }
 describe('NotificationContentDirective', () => {