+++ libs/core/src/lib/tile/tile-title/tile-title.directive.spec.ts
@@ -9,7 +9,7 @@ import { TileTitleDirective } from './tile-title.directive';
     `
 })
 class TestComponent {
-    @ViewChild('directiveElement', { static: false })
+    @ViewChild('directiveElement')
     ref: ElementRef;
 }