+++ libs/core/src/lib/token/public_api.ts
@@ -1,3 +1,4 @@
 export * from './token.module';
 export * from './token.component';
 export * from './tokenizer.component';
+export * from './token-input.directive';